refactor: dlt, package structure
work in progress, more changes are coming
