+++ src/main/kotlin/com/rootsid/wal/library/prism/model/DltDidUpdate.kt
@@ -1,0 +1,9 @@
+package com.rootsid.wal.library.prism.model
+
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class DltDidUpdate(
+    val operationId: String?,
+    val did: Did?
+)
